refactor(Dictionary): ディレクトリに移動
