+++ packages/@proofdict/editor/package.json
@@ -22,6 +22,7 @@
     "almin-logger": "^6.1.5",
     "almin-react-container": "^0.6.4",
     "classnames": "^2.2.5",
+    "ddd-base": "^0.3.1",
     "js-yaml": "^3.9.1",
     "kuromojin": "^1.3.2",
     "lodash.intersectionby": "^4.7.0",
