+++ packages/@proofdict/editor/src/domain/Dictionary/DictionaryDescription.ts
