+++ packages/@proofdict/editor/src/domain/Dictionary/DictionaryFactory.ts
