+++ packages/@proofdict/editor/src/domain/Dictionary/DictionaryPattern.ts
