+++ packages/@proofdict/editor/src/domain/Dictionary/DictionarySpecs.ts
@@ -1,19 +1,22 @@
 // MIT © 2017 azu
 import { splice } from "@immutable-array/prototype";
 import { DictionarySpec, DictionarySpecJSON } from "./DictionarySpec";
-import { Serializer } from "../ddd-base/Serializer";
+import { Serializer } from "../../ddd-base/Serializer";
 
 export const DictionarySpecsSerializer: Serializer<DictionarySpecs, DictionarySpecsJSON> = {
     fromJSON(specs) {
         return new DictionarySpecs(specs.map(spec => new DictionarySpec(spec)));
     },
     toJSON(entity) {
-        return entity.getSpecList().filter(spec => spec.isFilled).map(spec => {
-            return {
-                from: spec.from,
-                to: spec.to!
-            };
-        });
+        return entity
+            .getSpecList()
+            .filter(spec => spec.isFilled)
+            .map(spec => {
+                return {
+                    from: spec.from,
+                    to: spec.to!
+                };
+            });
     }
 };
