+++ packages/@proofdict/editor/src/domain/Dictionary/DictionaryWordClass.ts
