+++ packages/@proofdict/editor/src/infra/formatter/YamlFormatter.ts
@@ -1,6 +1,6 @@
 // MIT © 2017 azu
 import * as yaml from "js-yaml";
-import { DictionaryJSON } from "../../domain/Dictionary";
+import { DictionaryJSON } from "../../domain/Dictionary/Dictionary";
 
 /**
  * Output as Yaml String
