+++ packages/@proofdict/editor/src/infra/prh/Prh.ts
@@ -1,7 +1,7 @@
 // MIT © 2017 azu
 import { Engine } from "prh";
-import { Dictionary, DictionarySerializer } from "../../domain/Dictionary";
-import { DictionarySpec } from "../../domain/DictionarySpec";
+import { Dictionary, DictionarySerializer } from "../../domain/Dictionary/Dictionary";
+import { DictionarySpec } from "../../domain/Dictionary/DictionarySpec";
 import { ProofdictTester } from "proofdict-tester";
 
 const intersectionBy = require("lodash.intersectionby");
