+++ packages/@proofdict/editor/src/infra/repository/DictionaryRepository.ts
@@ -1,8 +1,8 @@
 // MIT © 2017 azu
-import { Dictionary } from "../../domain/Dictionary";
-import { NonNullableBaseRepository } from "../../ddd-base/NonNullableBaseRepository";
-import { createDictionary } from "../../domain/DictionaryFactory";
+import { Dictionary } from "../../domain/Dictionary/Dictionary";
+import { NonNullableRepository } from "ddd-base";
+import { createDictionary } from "../../domain/Dictionary/DictionaryFactory";
 
-export class DictionaryRepository extends NonNullableBaseRepository<Dictionary> {}
+export class DictionaryRepository extends NonNullableRepository<Dictionary> {}
 
 export const dictionaryRepository = new DictionaryRepository(createDictionary());
