+++ packages/@proofdict/editor/src/use-case/dictionary/AddNewPatternToDictionaryUseCase.ts
@@ -1,8 +1,8 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
-import { DictionaryIdentifier } from "../../domain/Dictionary";
-import { DictionaryPattern } from "../../domain/DictionaryPattern";
+import { DictionaryIdentifier } from "../../domain/Dictionary/Dictionary";
+import { DictionaryPattern } from "../../domain/Dictionary/DictionaryPattern";
 
 export const createAddNewPatternToDictionaryUseCase = () => {
     return new AddNewPatternToDictionaryUseCase({
