+++ packages/@proofdict/editor/src/use-case/dictionary/CreateNewDictionaryUseCase.ts
@@ -1,10 +1,10 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
-import { createDictionary } from "../../domain/DictionaryFactory";
-import { DictionaryPattern } from "../../domain/DictionaryPattern";
-import { DictionaryExpected } from "../../domain/DictionaryExpected";
-import { DictionarySpec } from "../../domain/DictionarySpec";
+import { createDictionary } from "../../domain/Dictionary/DictionaryFactory";
+import { DictionaryPattern } from "../../domain/Dictionary/DictionaryPattern";
+import { DictionaryExpected } from "../../domain/Dictionary/DictionaryExpected";
+import { DictionarySpec } from "../../domain/Dictionary/DictionarySpec";
 
 export const createCreateNewDictionaryUseCase = () => {
     return new CreateNewDictionaryUseCase({ dictionaryRepository });
