+++ packages/@proofdict/editor/src/use-case/dictionary/ResetDictionaryUseCase.ts
@@ -1,8 +1,8 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
-import { DictionaryIdentifier } from "../../domain/Dictionary";
-import { resetDictionary } from "../../domain/DictionaryFactory";
+import { DictionaryIdentifier } from "../../domain/Dictionary/Dictionary";
+import { resetDictionary } from "../../domain/Dictionary/DictionaryFactory";
 
 export const createResetDictionaryUseCase = () => {
     return new ResetDictionaryUseCase({
