+++ packages/@proofdict/editor/src/use-case/dictionary/UpdateDictionarySpecUseCase.ts
@@ -1,8 +1,8 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
-import { DictionaryIdentifier } from "../../domain/Dictionary";
-import { DictionarySpec } from "../../domain/DictionarySpec";
+import { DictionaryIdentifier } from "../../domain/Dictionary/Dictionary";
+import { DictionarySpec } from "../../domain/Dictionary/DictionarySpec";
 import { createUpdateDictionarySpecStatusUseCase } from "./UpdateDictionarySpecStatusUseCase";
 
 export const createUpdateDictionarySpecUseCase = () => {
